Angularised Assign user to issue
Assign a user to an issue" auto suggest has now user avatar’s with the
name.
Fixed a bug in assign user to issue autosuggest where the autosuggest
was not working after adding a member to issue.
Some UI improvements in the issues page. Now the selected iteration tab
is highlighted.